Maven Restructure + README Update
