+++ Dark-Heresy-2e/Character/pom.xml
@@ -4,12 +4,12 @@
 
 	<parent>
 		<groupId>nz.co.macro303.rpg-creation</groupId>
-		<artifactId>dungeons-and-dragons-5e</artifactId>
-		<version>0.0.5</version>
+		<artifactId>dark-heresy-2e</artifactId>
+		<version>0.0.6</version>
 	</parent>
 
-	<groupId>nz.co.macro303.rpg-creation.dungeons-and-dragons-5e</groupId>
-	<artifactId>characters</artifactId>
+	<groupId>nz.co.macro303.rpg-creation.dark-heresy-2e</groupId>
+	<artifactId>character</artifactId>
 	<packaging>jar</packaging>
-	<name>Characters</name>
+	<name>Character</name>
 </project>
