+++ Dark-Heresy-2e/Character/src/main/kotlin/macro303/character/HairColour.kt
@@ -1,4 +1,4 @@
-package macro303.dh2e
+package macro303.character
 
 enum class HairColour(val value: String) {
 	RED(value = "Red"),
