+++ Dark-Heresy-2e/Character/src/main/kotlin/macro303/character/Homeworld.kt
@@ -1,4 +1,4 @@
-package macro303.dh2e
+package macro303.character
 
 enum class Homeworld(val value: String) {
 	FERAL_WORLD(value = "Feral World"),
