--- Dark-Heresy-2e/Characters/src/main/kotlin/macro303/dh2e/DH2e.kt
@@ -1,17 +1,0 @@
-package macro303.dh2e
-
-object DH2e {
-
-	@JvmStatic
-	fun main(args: Array<String>) {
-		val total = 3
-		(0 until total).forEach {
-			val character = Builder.buildCharacter()
-			if (total <= 3) {
-				character.display()
-				println("=".repeat(50))
-			} else
-				println("Character: $character")
-		}
-	}
-}
