+++ Dungeons-and-Dragons-5e/Character/src/main/kotlin/macro303/character/Builder.kt
@@ -1,4 +1,4 @@
-package macro303.dnd5e
+package macro303.character
 
 import java.util.*
