--- Dungeons-and-Dragons-5e/Characters/src/main/kotlin/macro303/dnd5e/DnD5e.kt
@@ -1,17 +1,0 @@
-package macro303.dnd5e
-
-object DnD5e {
-
-	@JvmStatic
-	fun main(args: Array<String>) {
-		val total = 3
-		(0 until total).forEach {
-			val character = Builder.buildCharacter()
-			if (total <= 3) {
-				character.display()
-				println("=".repeat(50))
-			} else
-				println("Character: $character")
-		}
-	}
-}
